[PATCH] me/lib/mem_ring.c: Explicitly cast register pointer using declspec
This patch is for correctly dealing with current and foreseeable compiler
behavior. The need for this patch is fairly well documented in this JIRA:
https://jira.netronome.com/browse/THSDK-1366

Basically, we need to explicitly cast the 32bit register pointer inside an
intrinsic function as such, in order to avoid the compiler assuming a 40bit
cast and thus making the intrinsic function sort of return a 40 bit result.
Thats my understanding of it in any case. Seems this intrinsic function idea is
a rather loose concept in compiler land.

Signed-off-by: Mike Aitken <mike.aitken@netronome.com>
Acked-by: Rolf Neugebauer <rolf.neugebauer@netronome.com>