[PATCH 1/3] me/lib/pkt: Add basic packet manipulation utilities
This patch introduces structures and functions for receiving, manipulating and
sending packets in the NFP 6xxx series of processors. The functions should
provide the ability for the user to handle build a basic wire application as
well as modify the contents of the packets they are transporting.  These
changes do not provide access to all of the NFP MU packet engine operations;
just the basics.  Subsequent commits should build upon these operations.

I have tested these with a sample wire application under the Linux environment.
Both the wire application and the configuration tools required to bootstrap the
environment will follow in subsequent patches.

Signed-off-by: Christopher Telfer <chris.telfer@netronome.com>
Acked-by: Jon Hickman <jon.hickman@netronome.com>