[PATCH 3/3] me/apps/wire: Initial commit of a basic wire application in microC
This is a basic wire application for the NFP 6xxx written in microC
along with a shell script to run it in the Linux environment.  This wire
application does not modify packets in any way (except to prepend the
required traffic modifier rewrite script).  It is currently configured
to run on 2 MEs.  It does not recycle MU buffers.  Once the BLM code is
integrated into flowenv, this application should be modified make use of
it.

Building and running:
  - set the SDKHOME environment variable to refer to the top level SDK
    directory
  - set LD_LIBRARY_PATH to include $SDKHOME/lib
  - type 'make' in contrib/dbgtools to build the required sim config
    tools
  - type 'make' in me/apps/wire to build the microcode
  - start the simulator in one terminal by running 'nfsim'
  - type 'scripts/run_wire_sim.sh file1.pcap file2.pcap' in another
    terminal to run the wire app sending in file1.pcap into NBI0 port 0
    and file2.pcap into NBI1 port 0.
  - one can inspect the outputted pcaps in /tmp/wire_nbi0.pcap and
    /tmp/wire_nbi1.pcap

Signed-off-by: Christopher Telfer <chris.telfer@netronome.com>
Acked-by: Jon Hickman <jon.hickman@netronome.com>