contrib/simtools: Initial commit of Linux userspace sim tools
This patch introduces very basic tools for manipulating the NFP 6xxx
simulation environment from the Linux command line.  Many of these tools
are revised versions of example code provided with the SDK extended for
more flexibility.  Using these tools, one should be able to configure,
run and debug a basic wire application setup from the linux command line
in the flowenv environment.

I have tested this with a wire application that will be committed in a
subsequent patch.

Signed-off-by: Christopher Telfer <chris.telfer@netronome.com>
Acked-by: Rolf Neugebauer <rolf.neugebauer@netronome.com>