libstd: Add support for 4 ctx xfer numbers in event library.
The change in this patch is to handle formulating the CLS signal data
reference when in 8 or 4 ctx mode, it presently only works in 8 ctx mode.

The CLS signal data reference is made up of ctx+xfer and can be thought of a
as 8 bit value for the general purpose of explaining of how it is constructed.
Please refer to the Databook for further details.

When in 8 ctx mode the CLS signal data reference has 8 ctx (0 - 7) represented
by 3 bits and 32 transfer registers (0 - 31) represented by 5 bits.  CCCT TTTT

When in 4 ctx mode the CLS signal data reference has 4 ctx (0, 2, 4, 6)
represented by 2 bits and 64 transfer registers (0 - 63) represented by 6
bits. CCTT TTTT

So when in 8 ctx mode we use the msb bit from the transfer register to give
to ctx and in 4 ctx mode we use the lsb bit from the ctx to give to the
transfer register.

The ctx mask of 0x7 does not have to change for 4 ctx mode because the ctx
values in 4 ctx mode never have the lsb set the masks (0, 2, 4, 6). The
transfer register mask must be 0x3f. That mask will works for both 8 and 4 ctx
mode because in 8 ctx mode the transfer registers will never have the msb set
(0 - 31).

Signed-off-by: Jon Hickman <jon.hickman@netronome.com>
Signed-off-by: Michael Rapson <michael.rapson@netronome.com>
Acked-by: Rolf Neugebauer <rolf.neugebauer@netronome.com>