[init] Allow CTM offset to be configurable
The CTM offset can be 1 = 128B or 0 = 64B for B0 and 1 = 32B or 0 = 64B for A0.
This patch will allow for CTM offset to be overridden by a define value of
PKT_NBI_OFFSET.  If not defined it will use 64B as it did by default before.
If PKT_NBI_OFFSET is set to an invalid value or chip revision is not
recognized it will assert an error.

Signed-off-by: Jon Hickman <jon.hickman@netronome.com>
Acked-by: Sridhar Pinnapa <sridhar.pinnapa@netronome.com>
Acked-by: Sam Hong <samuel.hong@netronome.com>