pktio: fix compute_ctm_size() for a certain packet length range
The compute_ctm_size() function in libpktio.c as currently implemented,
computes the wrong value for ctm-sizes, in the case where the length of the
packet is 0x200-0x2FF.  It computes a "3" when it should be a "2".  For the
range 300-3FF, it correctly computes a "2".  The following patch fixes this
issue. Note that, surprisingly, this doesn't seem to break much; mainly
because most application code doesn't mess with the packet data beyond the
end of the CTM buffer.  But one exception is IPsec, where (all of) the packet
data gets encrypted/decrypted, then copied back out to the original "split"
data buffers.

Signed-off by: Matt Cimbala <matt.cimbala@sophos.com>
Acked by: Sam Hong <samuel.hong@netronome.com>