pktio: use NBI TM drop
This commit adds support for pktio to have TM do return of CTM and MU handles.
To enable define PKTIO_NBI_DROP_TXQ to a value in the range 0-1023,
corresponding to the TM queue drop counter which will reflect the drop.

Enabling this saves on two data-less IOs and around 30 instructions. This is
feature is not present on 6XXX chips.

Signed-off-by: David George <david.george@netronome.com>
Acked-by: Michael Rapson <michael.rapson@netronome.com>