me/lib/std: Offer pkt_cntr_atmc_xyz() API, compatible with pkt_cntr_xyz()
The pkt_cntr_xyz() API uses the stats engine to collect packet and byte
counters in a single op.  This is an efficient API to use when the stats
engine is available.  However, there are various reasons why it might
not be available.  The stats engine might not be present on certain
platforms, the memory available in that MU might be needed for other
purposes, or a debugging effort might need to move memory symbols, etc.

This patch adds an equivalent API, pkt_cntr_atmc_xyz() offering the
same function types with the same parameter lists and the same basic
behaviour.  PKT_CNTR_ALLOC_SYM() can be used to allocate the atomic
counter memory, as the backing structures are the same size.
A helper PKT_CNTR_ATMC_GET_ADDR_SYM(_name) is added.

There are inevitably subtle differences between the functions
in terms of xfers, signals, and GPRs needed internally, and in behaviour
if the counters are not accumulated timeously.

Signed-off-by: Michael Rapson <michael.rapson@netronome.com>
Acked-by: David George <david.george@netronome.com>