me/lib/std: add pkt_cntr_read_and_clr usage example and rate estimate
The division between bits assigned for packets and bytes in the stats
engine was selected to ensure that the packet count will not wrap
before the byte counts, assuming the minimum packet size is 64B.
That is, (1 << 29) * 64 = (1 << 35).  This means that we can calculate
accumulation rates based on worst case data rates only.

Technically we do receive packets smaller than 64B (minimum size packets
are 60B after removing FCS, and ARP requests can be smaller before sending
on the wire).  Nevertheless using data rates only is still justified as
the effective bandwidth for these sizes is lower as well.

If our worst case data rate is G (Gbps), the time to wrap in seconds is:
S = ((1 << 35) * 8) / (G * 1e9) = 274.87 / G.

Add some text to the header file with this approximation.

Signed-off-by: Michael Rapson <michael.rapson@netronome.com>
Acked-by: David George <david.george@netronome.com>