pktio: support ctm size encoded in nbi metadata
38xx platforms have the ctm size encoded in the nbi metadata. This means
there is no need to calculate it on RX and also means there is no need to
store the value in a separate field in the packet metadata. In order
to facilitate this, accessor functions pktio_ctm_size_get() and _set() are
introduced. 6xxx code will still work using the pkt.p_ctm_size field, but
38xx code will need to change over to the accessor functions.

Signed-off-by: David George <david.george@netronome.com>
Acked-by: Michael Rapson <michael.rapson@netronome.com>