me/lib/nfp: Update struct pcie_dma_cfg_one for NFP3800
The NFP3800 brings changes to "struct nfp_pcie_dma_cfg" to
support CTM free on last and FromPCI CPP round8 features.
That struct is provided via nfp_pcie.h, so flowenv gets the
updated definitions from the autogenerated header files.

Flowenv maintains "struct pcie_dma_cfg_one" in pcie.h as part
of an API that allows users to set either the even or odd config
in the pair.  This must be updated by hand.

Future chips are likely to make other changes as features are
added and removed, so use an #if-#elif-#else to ensure this
struct is checked early in the future.

Signed-off-by: Michael Rapson <michael.rapson@netronome.com>
Acked-by: David George <david.george@netronome.com>