blm: Fix enable mask unreachable code warning
When the BLM_BLQ_ENABLE_MASK is used to disable contexts, the
BLM generates unreachable code errors. This is because a runtime
.if/.else/.endif test is used to handle the blm_is_blq_enable()
test, but the assembler can detect that ".if" is not taken.
Simply changing this to a #if/#else/#endif still causes warnings
as the code structure leaves code below the ctx_arb[kill].

Instead, test the value in a register so that the assembler
does not emit the warnings.  This uses extra codestore words,
so test for the common case, BLM_BLQ_ENABLE_MASK 0xff, explicitly
and handle it specially.

Also ensure that contexts used for PCIe are covered by the
enabled test.

Signed-off-by: Michael Rapson <michael.rapson@netronome.com>
Acked-by: Sean Pennefather <sean.pennefather@netronome.com>