blm: Ensure consistent use of BLM_BLQ_ENABLE_MASK
Currently blm_main.uc interprets bits as 0..3 for ingress
and 4..7 for egress, but blm_cfg.h assumes each bit represents
the particular context.  These need to be consistent for
BLM_BLQ_ENABLE_MASK to be used correctly.

Having each bit represent enabling/disabling the associated
context is one less layer of abstraction and is probably
more straightforward for developers with access to the code,
so let's opt for that interpretation.

Signed-off-by: Michael Rapson <michael.rapson@netronome.com>
Acked-by: Sean Pennefather <sean.pennefather@netronome.com>