init: Handle PCI DMA init
Configure the various PCIe DMA related CSRs that are not handled
by the PCIe BLM, see init_pci_dma.uc for a breakdown of the CSRs.

Introduce PCIe specific defines: PCI_DMA_SPLIT_LENGTH,
PCI_DMA_CTM_OFFSET, and PCI_DMA_BLQ_BYPASS.  To support
BLQ bypass, also add a global define PKT_NULL_MU_PTR and
ensure that this value is configured correctly in the NBI and
CTMs.

Buffer pools are configured in a similar way to NBI, via
PCI_DMA_BPx_BLQ_TARGET defines.  The NFP38xx supports new
buffer padding options for the BLQs.  For now expose this
via the PCI_DMA_CTM_BUF_PAD option.

Enable CTM polling from the DMA engine.  This should be disabled
by soft reset.  The general recommendation is to set this _after_
the reset of the init is done.  However, the NBI seems to do it
during the init phase without issue, so copy that for the PCI.

Signed-off-by: Michael Rapson <michael.rapson@netronome.com>
Acked-by: Sean Pennefather <sean.pennefather@netronome.com>